Imputation by Cummutative Functions Impute using replacement values calculated from a univariate, cummuative function.

Imputation by Cummutative Functions

Impute using replacement values calculated from a univariate, cummuative function. na.median imputes with the median value of x. The median is only valid for numeric or logical values.

na.max(.x, ...)

na.min(.x, ...)

na.mean(.x, ...)

na.median(.x, ...)

na.quantile(.x, ...)

na.most_freq(.x, ...)

Arguments

.x
vector in which NA values are to be replaced. The ordering of x does not matter.

...
additional arguments passed to lower-level summary functions.

Value

A vector of class(x) and length(x) in which missing values (NA) have been replaced the result of a function call: $$ fun(x, ...) $$

Details

This collection of functions calculates a replacement value using an unvariate function where the order of values in x do not matter, i.e. cummutative. na.max and na.min replace missing values (NA) with the maximum or minimum of non-missing values x. ( Internally: base::max(..., na.rm=TRUE) and base::min(..., na.rm=TRUE). ... has no affect. na.mean replaces NA values with the mean of x. Internally, mean(x, na.rm=TRUE, ... ) is used. If mean cannot be calculated (e.g. x isn't numeric) then x is returned with a warning. na.quantile imputes with a quantile given by the prob argument that is passed to quantile() na.most_freq replaces all NA with the most frequently occuring value. In the event of ties, the value encounter first is listed first is used.

See also

na.replace() - used internally by these functions

na.constant()

base::max() and base::min()

median()

quantile()

Examples

  na.median( c(1,2,NA_real_,3) )

#> [1] 1 2 2 3

  na.quantile( c(1,2,NA_real_,3), prob=0.4 )

#> [1] 1 2 2 3

  c(1,1,NA,4) %>% na.most_freq()

#> [1] 1 1 1 4

  c(1,1,4,4,NA) %>% na.most_freq()

#> [1] 1 1 4 4 1