Replace Missing Values

Replaces NA values with explicit values.

na.replace(x, .na, ...)

na.explicit(x, .na)

Arguments

x
vector in which NA values are to be replaced.

.na
scalar, length(x)-vector or function used to replace NA. See #Details.

...
additional arguments passed to .na when it is a function.

Value

A vector with the class and length of x. NAs in x will be replaced by .na. .na is coerced as necessary.

Details

na.replace replaces missing values in x by .na if possible.

In R, replacement of values can cause the class/type of to change. This is not often desired behavior. By contrast, na.replace replaces values in class/type-safe and length-safe ways. The result is guaranteed to be the same class/type and length of x regardless of the value of .na. Param: x

If x is categorical (e.g. character or factor), .na is optional. The default is "(NA)" and can be set with options( NA_explicit_ = new_value ). It can also be referenced directly with NA_explicit_.

If x is a factor, unique values of .na not in already present in levels(x) will be added. They are appended silently unless getOption('verbose')==TRUE in which a message reports the added levels. Param: .na .na can be either a scalar, vector or function.

If a scalar, each missing value of x is replaced by na.

If a vector, .na must have length(x). Missing values ofxare replaced by corresponding elements of.na. Recycling values of.nais not performed. An error will be thrown in the event thatlength(.na)is not1orlength(x).

If a function, x is transformed by .na` with:

     .na( na.omit(x) )

then preceding with normal operations. na.explicit is an alias for na.replace that is likely to be deprecated.

See also

base::ifelse(), base::replace()

forcats::fct_explicit_na - which only handles factors

Examples

  # Integers and numerics
  na.replace( c(1,NA,3,NA), 2 )    # 1 2 3 2   

#> [1] 1 2 3 2

  na.replace( c(1,NA,3,NA), 1:4 )  # 1 2 3 4

#> [1] 1 2 3 4

  # This produces an error because it would change the type
  
# NOT RUN {
  na.replace( c(1,NA,3,NA), letters[1:4] )  # "1" "b" "3" "d"

# }

  # Characters 
  lets <- letters[1:5]
  lets[ c(2,4) ] <- NA
  na.replace(lets)  # replace with NA_explicit_

#> [1] "a"    "(NA)" "c"    "(NA)" "e"   

  # Factors 
  fct <- as.factor( c( NA, letters[2:4], NA) )
  fct

#> [1] <NA> b    c    d    <NA>
#> Levels: b c d

  na.replace(fct, "z")  # z b c d z  -- level z added

#> [1] z b c d z
#> Levels: b c d z

  na.replace(fct, letters[1:5] )

#> [1] a b c d e
#> Levels: b c d a e

  na.replace(fct)

#> [1] (NA) b    c    d    (NA)
#> Levels: b c d (NA)

     
 
# NOT RUN {
   na.replace( rep(NA,3), rep(NA,3) )

# }